+++ mq-jms-spring-boot-starter/src/main/java/com/ibm/mq/spring/boot/MQConfigurationPropertiesJks.java
@@ -53,7 +53,7 @@ public class MQConfigurationPropertiesJks {
   public void setTrustStore(String trustStore) {
     this.trustStore = trustStore;
   }
-
+  
   public String getKeyStorePassword() {
     return keyStorePassword;
   }